First version
This is a mess of in developement code. Everything is in the "frontend"
(which allows easy running of the server and the jsaddle frontend at the
same time).

The assets, used in Assets/.hs, are not part of the source until I'm not
creating my own version of it, mostly because right now I'm using scan
of the original game.