Huge cleanup of the cabal and tree hierarchy.
We use the same cabal file for the different libraries.

This is discutable because spliting in many librarys (common, frontend,
server) may lead to an easier to understand codebase, and probably a
smaller GHCJS output.

However it is really convenient because with one `ghci` in can load the
jsaddle-warp and server code at the same time and don't need to reload
the full project if anything changes.